# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Personal Finance API
  description: |
    API for the Personal Finance application. Provides endpoints for managing
    bank accounts, transactions, investments, market data, and notifications.

    ## Authentication
    Most endpoints require authentication via Supabase Auth. Include the session
    cookie or Authorization header with a valid JWT token.

    ## Response Format
    Successful responses follow this format:
    ```json
    {
      "success": true,
      "data": { ... },
      "meta": { "count": 10 }
    }
    ```

    Error responses follow this format:
    ```json
    {
      "success": false,
      "error": "Error message"
    }
    ```
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: /api
    description: API routes
...
paths:
...
  /teller/accounts:
    get:
      tags: [Accounts]
      summary: List connected bank accounts
      description: Returns all connected bank accounts for the authenticated user
      security:
        - cookieAuth: []
      responses:
        '200':
          description: List of accounts
          content:
            application/json:
              schema:
                type: object
                properties:
                  accounts:
                    type: array
                    items:
                      $ref: '#/components/schemas/BankAccount'
                  count:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      tags: [Accounts]
      summary: Disconnect a bank enrollment
      description: Soft deletes a bank enrollment and its associated accounts
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enrollmentId]
              properties:
                enrollmentId:
                  type: string
                  description: The enrollment ID to disconnect
      responses:
        '200':
          description: Successfully disconnected
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /market/quote:
    get:
      tags: [Market]
      summary: Get stock quote
      description: Fetches current price and data for a stock symbol
      security:
        - cookieAuth: []
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
          example: "AAPL"
      responses:
        '200':
          description: Stock quote data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TickerData'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /market/insights:
    get:
      tags: [Market]
      summary: Get market insights
      description: Returns market insights and trends
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Market insights
        '401':
          $ref: '#/components/responses/Unauthorized'
  /market/holdings:
    get:
      tags: [Market]
      summary: Get user holdings
      description: Returns user's investment holdings with current values
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Holdings list
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============ INVESTMENTS ============
...
  /notification-schedules:
    get:
      tags: [Discord]
      summary: Get notification schedules
      description: Returns user's notification schedule preferences
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Notification schedules
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Discord]
      summary: Update notification schedules
      description: Updates notification schedule preferences
      security:
        - cookieAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                schedules:
                  type: array
                  items:
                    type: object
                    properties:
                      type:
                        type: string
                        enum: [daily_summary, weekly_report, market_alerts, budget_alerts]
                      enabled:
                        type: boolean
                      time:
                        type: string
                      day:
                        type: string
      responses:
        '200':
          description: Schedules updated
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============ CHAT ============
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: Supabase session cookie
    cronAuth:
      type: http
      scheme: bearer
      description: CRON_SECRET bearer token for scheduled jobs
    discordSignature:
      type: apiKey
      in: header
      name: X-Signature-Ed25519
      description: Discord interaction signature
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: "Unauthorized"
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: "Invalid request"
  schemas:
...
    BankAccount:
      type: object
      properties:
        id:
          type: string
        accountId:
          type: string
        name:
          type: string
          example: "Checking Account"
        type:
          type: string
          enum: [depository, credit, loan, investment]
        subtype:
          type: string
          example: "checking"
        institutionName:
          type: string
          example: "Chase"
        lastFour:
          type: string
          example: "1234"
        currentBalance:
          type: number
          example: 1500.00
        availableBalance:
          type: number
          example: 1450.00
        creditLimit:
          type: number
          nullable: true
        currency:
          type: string
          example: "USD"
        lastSynced:
          type: string
          format: date-time
    TickerData:
      type: object
      properties:
        symbol:
          type: string
          example: "AAPL"
        name:
          type: string
          example: "Apple Inc"
        price:
          type: number
          example: 175.50
        change:
          type: number
          example: 2.35
        changePercent:
          type: number
          example: 1.36
        volume:
          type: integer
          example: 45000000
        marketCap:
          type: number
          example: 2800000000000
        performance30d:
          type: number
          example: 5.2
        performance90d:
          type: number
          example: 12.8
        performance365d:
          type: number
          example: 28.5
        timestamp:
          type: string
          format: date-time
...